fix: use long instead of int I realized that int uses 32bit, but using long uses 64bit, that should fix it but oh well.

+++ StoatApplication/Core/WebSocket/Models/ClientToServer/Ping.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 namespace StoatApplication.Core.WebSocket.Models.ClientToServer;
 
 public sealed record Ping(
-    [property: JsonPropertyName("data")] int Data
+    [property: JsonPropertyName("data")] long Data
 ) : IEvent
 {
     public string Type => "Ping";
